published new versions (patch) of packages: @bemoje/commander-config@0.2.31, @bemoje/repoman@0.0.45, @bemoje/terminal-gpt@0.2.36, @bemoje/util@0.0.28

+++ packages/repoman/src/lib/deletePackage.ts
@@ -7,8 +7,10 @@ export function deletePackage(name: string) {
   if (!name) throw new Error('No name provided')
 
   execute(`npm uninstall @bemoje/${name}`)
-  getPackages().forEach(({ rootdir }) => {
-    execute(`npm uninstall @bemoje/${name}`, { cwd: rootdir })
+  getPackages().forEach(({ rootdir, pkg }) => {
+    if (Object.keys(pkg.dependencies || {}).includes(`@bemoje/${name}`)) {
+      execute(`npm uninstall @bemoje/${name}`, { cwd: rootdir })
+    }
   })
   execute(`nx g @nrwl/workspace:remove ${name}`)
